% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_bulk.R
\name{as_BulkPhyloExpressionSet}
\alias{as_BulkPhyloExpressionSet}
\title{as_BulkPhyloExpressionSet}
\usage{
as_BulkPhyloExpressionSet(
  data,
  groups = colnames(data[, 3:ncol(data)]),
  name = deparse(substitute(data)),
  strata_legend = NULL,
  ...
)
}
\description{
This function is an alias for \code{BulkPhyloExpressionSet_from_df}. 
Please refer to the documentation for \code{BulkPhyloExpressionSet_from_df} for usage details, arguments, and examples.
}
\seealso{
\code{\link{BulkPhyloExpressionSet_from_df}}
}
