% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{available_identities}
\alias{available_identities}
\title{Get available identities}
\usage{
available_identities(phyex_set)
}
\arguments{
\item{phyex_set}{A ScPhyloExpressionSet object}
}
\value{
Character vector of available identity columns
}
\description{
Return the available metadata columns that can be used as identities in the Seurat object.
}
