% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bar.colors}
\alias{bar.colors}
\title{A function to get a vector of length n storing a palette of colors for multiple bars in barplots}
\usage{
bar.colors(n)
}
\arguments{
\item{n}{the number of colors to be in the palette.}
}
\value{
a character vector containing different color names that can be used for barplots.
}
\description{
A nice color palette for barplots with several bars.
}
\examples{
# get 5 different colors for 5 different bars
barplot_colors <- bar.colors(5)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{palette}}
}

