% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_bulk.R
\name{BulkPhyloExpressionSet}
\alias{BulkPhyloExpressionSet}
\title{Bulk PhyloExpressionSet Class}
\usage{
BulkPhyloExpressionSet(
  strata = stop("@strata is required"),
  gene_ids = stop("@gene_ids is required"),
  name = "Phylo Expression Set",
  species = character(0),
  index_type = "TXI",
  identities_label = "Identities",
  null_conservation_sample_size = 5000L,
  precomputed_null_conservation_txis = NULL,
  expression = stop("@expression is required")
)
}
\arguments{
\item{expression}{Matrix of expression counts with genes as rows and samples as columns}

\item{groups}{Factor vector indicating which group/condition each sample belongs to}
}
\description{
S7 class for bulk RNA-seq phylotranscriptomic expression data.
This class handles expression data with biological replicates.
}
