% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{check_PhyloExpressionSet}
\alias{check_PhyloExpressionSet}
\title{Check if object is a PhyloExpressionSet}
\usage{
check_PhyloExpressionSet(phyex_set)
}
\arguments{
\item{phyex_set}{An object to check}
}
\value{
Invisibly returns TRUE if check passes, otherwise throws an error
}
\description{
Checks if the input is a PhyloExpressionSet S7 object and throws an error if not.
}
