% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{check_ScPhyloExpressionSet}
\alias{check_ScPhyloExpressionSet}
\title{Check if object is a ScPhyloExpressionSet}
\usage{
check_ScPhyloExpressionSet(phyex_set)
}
\arguments{
\item{phyex_set}{An object to check}
}
\value{
Invisibly returns TRUE if check passes, otherwise throws an error
}
\description{
Checks if the input is a PhyloExpressionSet S7 object and throws an error if not.
}
