% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{collapse}
\alias{collapse}
\title{Collapse PhyloExpressionSet Replicates}
\usage{
collapse(phyex_set, ...)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{...}{Additional arguments passed to methods}
}
\value{
A new PhyloExpressionSet object with collapsed expression data
}
\description{
Convert a PhyloExpressionSet with replicates to one with collapsed expression data.
}
\examples{
# Collapse replicates in a PhyloExpressionSet
collapsed_set <- collapse(example_phyex_set)

}
