% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_distributions.R
\docType{data}
\name{distributions}
\alias{distributions}
\title{Predefined Distribution Objects}
\format{
A named list containing Distribution objects:
\describe{
  \item{normal}{Normal distribution with mean and sd parameters}
  \item{gamma}{Gamma distribution with shape and rate parameters}
}
}
\usage{
distributions
}
\description{
List of predefined Distribution objects for use in statistical testing.
}
\details{
This list provides ready-to-use Distribution objects for common statistical
tests. Each distribution includes appropriate fitting functions and statistical
functions for hypothesis testing.
}
\examples{
# Use normal distribution for testing
# test_result <- generic_conservation_test(phyex_set, 
#                                         fitting_dist = distributions$normal)

}
\keyword{datasets}
