% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivergenceStratigraphy.R
\name{DivergenceStratigraphy}
\alias{DivergenceStratigraphy}
\title{Perform Divergence Stratigraphy}
\usage{
DivergenceStratigraphy(query, subject, subject.id = FALSE, eval = "1E-5",
  ortho.inference = "RBH", blast.path = NULL, cores = 1,
  dnds.threshold = 2, quiet = FALSE, clean = FALSE, ds.values = TRUE)
}
\arguments{
\item{query}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{subject.id}{a logical value indicating whether \code{query_id} AND \code{subject_id} should be returned as part of the output \code{Sequence Divergence Map}.}

\item{eval}{a numeric value specifying the e-value cutoff for BLASTP hit detection.}

\item{ortho.inference}{a character string specifying the orthology inference method that shall be performed
to detect orthologous genes. Default is \code{ortho.inference} = "RBH" (BLAST reciprocal best hit).
Available methods are: "BH" (BLAST best hit), "RBH" (BLAST reciprocal best hit).}

\item{blast.path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{cores}{a numeric value specifying the number of cores that shall be used to perform
parallel computations on a multicore machine.}

\item{dnds.threshold}{a numeric value specifying the dnds threshold for genes that shall be retained.
Hence all genes having a dNdS value <= \code{dnds.threshold} are retained. Default is \code{dnds.threshold} = 2.}

\item{quiet}{a logical value specifying whether a successful interface call to the \code{blastp} algorithm shall be printed out to the console.}

\item{clean}{a logical value specifying whether the internal folder structure shall be deleted (cleaned) after
processing this function. Default is \code{clean_folders} = \code{FALSE}. Hence, all BLAST results, pairwise alignments, and dNdS computation results stored in the \code{\link{tempdir}} folder will be removed and only the \code{Sequence Divergence Map} is returned.}

\item{ds.values}{a logical value specifying whether divegrence stratum values (ds values) or dNdS values shall be returned
by \code{divergence_stratigraphy}. Default is \code{ds.values} = \code{TRUE}.}
}
\value{
A data.table storing the Sequence Divergence Map of the query organism.
}
\description{
This function takes the CDS file of a query organism and performs
the Divergence Stratigraphy algorithm (Quint et al.,2012 ; Drost et al. 2015) to construct a 
Sequence Divergence Map for the corresponding query organism against a
closely related subject organism.
}
\details{
Introduced by Quint et al., 2012 and extended by Drost et al. 2015, Divergence Stratigraphy
 is an algorithm to quantify the selection pressure (in terms of amino acid sequence divergence) acting on
 orthologous genes between closely related species. The resulting Sequence Divergence Map (short divergence map),
 stores the divergence stratum assignment of each protein coding gene in the first column and the query_id of inferred orthologous genes in the second column.
 
 In brief, the Divergence Stratigraphy Algorithm can be summarized by the following procedure:
  
 \itemize{
 \item 1) Perform orthology Inference using BLAST reciprocal best hit ("RBH") based on the \code{blastp} algorithm
 
 \item 2) Perform pairwise global amino acid alignments of orthologous genes using the Needleman-Wunsch algorithm
 
 \item 3) Perform codon alignments of orthologous genes using PAL2NAL
 
 \item 4) dNdS estimation using Comeron's method (1995)
 
 \item 5) Assigning dNdS values to divergence strata (deciles of the dNdS distribution)
}
}
\note{
Although this function has been heavily optimized and parallelized, performing
Divergence Stratigraphy using two genomes will take some computation time.

In our experience performing Divergence Stratigraphy using two genomes (one query and one subject genome)
on an 8 core machine can take up to 1,5 - 2 hours.
}
\examples{
\dontrun{
 
 
 # Performing standard Divergence Stratigraphy and obtain a Sequence Divergence Map = SDM
 SDM <- DivergenceStratigraphy(
            query           = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject         = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
            eval            = "1E-5",
            ortho.inference = "RBH", 
            cores           = 1, 
            quiet           = TRUE, 
            clean           = TRUE)
      
      
      
 # performing standard Divergence Stratigraphy using the 'blast.path' argument to specify
 # the path to the blastp algorithm
 SDM <- DivergenceStratigraphy(
            query           = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject         = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
            eval            = "1E-5", 
            ortho.inference = "RBH",
            blast.path      = "path/to/blastp",
            cores           = 1, 
            quiet           = TRUE, 
            clean           = TRUE)
 
 
 
 # Divergence Stratigraphy can also be performed in parallel 
 # (on a multicore machine) by specifying the 'cores' argument
 SDM <- DivergenceStratigraphy(
            query           = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject         = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
            eval            = "1E-5", 
            ortho.inference = "RBH", 
            cores           = 2, 
            quiet           = TRUE, 
            clean           = TRUE)
 
 
 
 
 # in case users wish to generate a divergence map with DS values and 'query_id' and 'subject_id'
 # they can specify 'subject.id = TRUE'
 # performing standard Divergence Stratigraphy
 SDM <- DivergenceStratigraphy(
            query           = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject         = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
            subject.id      = TRUE,
            eval            = "1E-5",
            ortho.inference = "RBH", 
            cores           = 1, 
            quiet           = TRUE, 
            clean           = TRUE)
      
      
      
      
 # in case users wish to generate a divergence map with KaKs values and 'query_id' and 'subject_id'
 # they can specify 'ds.values = FALSE' and 'subject.id = TRUE'
 SDM <- DivergenceStratigraphy(
            query           = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject         = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
            subject.id      = TRUE,
            eval            = "1E-5",
            ortho.inference = "RBH", 
            cores           = 1, 
            quiet           = TRUE, 
            clean           = TRUE,
            ds.values       = FALSE)
            
            
            
# Perform the same analysis with Blast best hit as 'ortho.inference' method
SDM <- DivergenceStratigraphy(
            query           = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject         = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
            subject.id      = TRUE,
            eval            = "1E-5",
            ortho.inference = "BH", 
            cores           = 1, 
            quiet           = TRUE, 
            clean           = TRUE,
            ds.values       = FALSE)            
 }
 
}
\author{
Hajk-Georg Drost
}
\references{
Quint M et al. (2012). A transcriptomic hourglass in plant embryogenesis. Nature (490): 98-101.
 
 Drost HG et al. (2015). Evidence for Active Maintenance of Phylotranscriptomic Hourglass Patterns in Animal and Plant Embryogenesis. Mol Biol Evol. 32 (5): 1221-1231 doi:10.1093/molbev/msv012
}
\seealso{
\code{\link{PlotPattern}}, \code{\link{TDI}},  \code{\link{PlotRE}}, \code{\link{RE}}
}

