% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_distributions.R
\name{.fit_normal}
\alias{.fit_normal}
\title{Fit Normal Distribution Parameters}
\usage{
.fit_normal(x)
}
\arguments{
\item{x}{Numeric vector of data values}
}
\value{
List with mean and sd parameters
}
\description{
Fit normal distribution parameters using method of moments.
}
\examples{
# Fit normal distribution
# params <- .fit_normal(data_vector)

}
\keyword{internal}
