% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.get_cell_metadata}
\alias{.get_cell_metadata}
\title{Get Cell Metadata}
\usage{
.get_cell_metadata(seurat, cell_identity)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cell_identity}{Column name for cell identity}
}
\value{
Data frame of cell metadata
}
\description{
Extract relevant cell metadata from Seurat object.
}
\keyword{internal}
