% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.get_expression_from_seurat}
\alias{.get_expression_from_seurat}
\title{Get Expression Matrix from Seurat Object}
\usage{
.get_expression_from_seurat(seurat, layer)
}
\arguments{
\item{seurat}{A Seurat object}

\item{layer}{Character string specifying which layer to use}
}
\value{
Expression matrix (sparse or dense)
}
\description{
Extract expression matrix from Seurat object, preserving sparse format when possible.
}
\keyword{internal}
