% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.get_valid_cell_types}
\alias{.get_valid_cell_types}
\title{Get Valid Cell Types}
\usage{
.get_valid_cell_types(seurat, cell_identity, min_cells)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cell_identity}{Column name in metadata}

\item{min_cells}{Minimum number of cells per type}
}
\value{
Factor of valid cell types
}
\description{
Get cell types that meet minimum cell count requirements.
}
\keyword{internal}
