% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.map_cells_to_groups}
\alias{.map_cells_to_groups}
\title{Map Cells to Groups}
\usage{
.map_cells_to_groups(seurat, cell_identity, valid_identities)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cell_identity}{Column name in metadata}

\item{valid_identities}{Valid identity levels}
}
\value{
Factor mapping cells to groups
}
\description{
Create factor mapping each cell to its group/identity.
}
\keyword{internal}
