% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_null_conservation_txis.R
\name{.memo_generate_conservation_txis}
\alias{.memo_generate_conservation_txis}
\title{Memoized Null Conservation TXI Generation}
\usage{
.memo_generate_conservation_txis(strata_vector, count_matrix, sample_size)
}
\description{
Memoized version of generate_conservation_txis for improved performance
with repeated calls using the same parameters.
}
\details{
This function caches results to avoid recomputing expensive permutations
when the same parameters are used multiple times.
}
\keyword{internal}
