% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.pseudobulk_expression}
\alias{.pseudobulk_expression}
\title{Create Pseudobulk Expression Data}
\usage{
.pseudobulk_expression(seurat, cell_identity, slot, min_cells)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cell_identity}{Column name in metadata}

\item{slot}{Expression slot to use}

\item{min_cells}{Minimum cells per type}
}
\value{
Matrix of pseudobulked expression
}
\description{
Aggregate single-cell data by cell type to create pseudobulk expression.
}
\keyword{internal}
