% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{.pTXI}
\alias{.pTXI}
\title{Calculate pTXI for Raw Expression Data}
\usage{
.pTXI(expression_matrix, strata)
}
\arguments{
\item{expression_matrix}{Matrix of expression values}

\item{strata}{Vector of phylostratum assignments}
}
\value{
Matrix of pTXI values
}
\description{
Internal function to calculate pTXI for expression data.
}
\keyword{internal}
