% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{.pTXI_raw}
\alias{.pTXI_raw}
\title{Calculate pTXI for Raw Expression Data}
\usage{
.pTXI_raw(expression_matrix, strata)
}
\arguments{
\item{expression_matrix}{Matrix of expression values}

\item{strata}{Vector of phylostratum assignments}
}
\value{
Matrix of pTXI values
}
\description{
Internal function to calculate pTXI for raw (non-collapsed) expression data.
Used for TXI_sample calculations.
}
\keyword{internal}
