% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{.TXI}
\alias{.TXI}
\title{Calculate TXI for Raw Expression Data}
\usage{
.TXI(expression_matrix, strata)
}
\arguments{
\item{expression_matrix}{Matrix of expression values}

\item{strata}{Vector of phylostratum assignments}
}
\value{
Vector of TXI values
}
\description{
Internal function to calculate TXI for expression data.
}
\keyword{internal}
