% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.TXI_sc}
\alias{.TXI_sc}
\title{Calculate TXI for single cell expression sparse matrix.}
\usage{
.TXI_sc(expression, strata)
}
\arguments{
\item{expression}{Matrix of expression values, dgmatrix}

\item{strata}{Vector of phylostratum assignments}
}
\value{
Vector of TXI values
}
\description{
Internal function to calculate TXI for expression data.
}
\keyword{internal}
