% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EarlyConservationTest}
\alias{EarlyConservationTest}
\title{Function to perform the Early Conservation Test.}
\usage{
EarlyConservationTest(ExpressionSet, modules = NULL, permutations = 1000,
  lillie.test = FALSE, plotHistogram = FALSE, parallel = FALSE,
  runs = 10)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{modules}{a list storing three elements: early, mid, and late. Each element expects a numeric
vector specifying the developmental stages or experiments that correspond to each module.
For example, \code{module} = list(early = 1:2, mid = 3:5, late = 6:7) devides a dataset
storing seven developmental stages into 3 modules.}

\item{permutations}{a numeric value specifying the number of permutations to be performed for the \code{ReductiveHourglassTest}.}

\item{lillie.test}{a boolean value specifying whether the Lilliefors Kolmogorov-Smirnov Test shall be performed to quantify the goodness of fit.}

\item{plotHistogram}{a boolean value specifying whether a \emph{Lillifor's Kolmogorov-Smirnov-Test}
shall be performed to test the goodness of fit of the approximated distribution, as well as additional plots quantifying the significance
of the observed phylotranscriptomic pattern.}

\item{parallel}{a boolean value specifying whether goodness of fit computations
shall be performed in parallel using the \pkg{doMC} package.}

\item{runs}{specify the number of runs to be performed for goodness of fit computations, in case \code{plotHistogram} = \code{TRUE}.
In most cases \code{runs} = 100 is a reasonable choice. Default is \code{runs} = 10 (because it takes less computation time for demonstration purposes).}
}
\value{
a list object containing the list elements:

p.value : the p-value quantifying the statistical significance (low-high-high pattern) of the given phylotranscriptomics pattern.

std.dev : the standard deviation of the N sampled phylotranscriptomics patterns for each developmental stage S.

lillie.test : a boolean value specifying whether the \emph{Lillifors KS-Test} returned a p-value > 0.05,
which indicates that fitting the permuted scores with a normal distribution seems plausible.
}
\description{
The \emph{Early Conservation Test} has been developed to statistically evaluate the
existence of a monotonically increasing phylotranscriptomic pattern based on \code{\link{TAI}} or \code{\link{TDI}} computations.
The corresponding p-value quantifies the probability that a given TAI or TDI pattern (or any phylotranscriptomics pattern)
does not follow an early conservation like pattern. A p-value < 0.05 indicates that the corresponding phylotranscriptomics pattern does
indeed follow an early conservation (low-high-high) shape.
}
\details{
The \emph{Early Conservation Test}
}
\examples{
\dontrun{

data(PhyloExpressionSetExample)
data(DivergenceExpressionSetExample)

# perform the early conservation test for a PhyloExpressionSet
# here the prior biological knowledge is that stages 1-2 correspond to module 1 = early,
# stages 3-5 to module 2 = mid (phylotypic module), and stages 6-7 correspond to
# module 3 = late
EarlyConservationTest(PhyloExpressionSetExample,
                       modules = list(early = 1:2, mid = 3:5, late = 6:7),
                       permutations = 1000)


# perform the early conservation test for a DivergenceExpressionSet
# here the prior biological knowledge is that stages 1-2 correspond to module 1 = early,
# stages 3-5 to module 2 = mid (phylotypic module), and stages 6-7 correspond to
# module 3 = late
EarlyConservationTest(DivergenceExpressionSetExample,
                       modules = list(early = 1:2, mid = 3:5, late = 6:7),
                       permutations = 1000)

# to test the goodness of fit and to plot the histogram and Cullen and Frey skewness-kurtosis plot
# the parameter plotHistogram needs to be set to TRUE
# in case plotHistogram = TRUE, the runs parameter must be specified (e.g. 10 or 100)
# to test for the influence of permutation bias on the p-value
# runs = 10 determines the number of independent permutation runs that are
# performed to compute 10 independent p-values based on 10 independent underlying permutations
# and to search for permutation runs that cross the p < 0.05 border
EarlyConservationTest(PhyloExpressionSetExample,
                       modules = list(early = 1:2, mid = 3:5, late = 6:7),
                       permutations = 1000, lillie.test = TRUE,
                       plotHistogram = TRUE, runs = 10)


# since 10 runs using 1000 permutations for each run is computationally expensive
# the parameter parallel = TRUE can be used to perform parallel (multicore) processing
EarlyConservationTest(PhyloExpressionSetExample,
                       modules = list(early = 1:2, mid = 3:5, late = 6:7),
                       permutations = 1000, lillie.test = TRUE,
                       plotHistogram = TRUE, parallel = TRUE, runs = 10)



}
}
\author{
Hajk-Georg Drost
}
\references{
Drost et al. (2014), Active maintenance of phylotranscriptomic hourglass patterns in animal and plant embryogenesis.

Quint M et al. (2012). "A transcriptomic hourglass in plant embryogenesis". Nature (490): 98-101.

Piasecka B, Lichocki P, Moretti S, et al. (2013) The hourglass and the early conservation models--co-existing
patterns of developmental constraints in vertebrates. PLoS Genet. 9(4): e1003476.
}
\seealso{
\code{\link{gpScore}}, \code{\link{bootMatrix}}, \code{\link{FlatLineTest}}, \code{\link{PlotPattern}}
}

