% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_phyex_set_old}
\alias{example_phyex_set_old}
\title{Example phyex set old}
\format{
A BulkPhyloExpressionObject with 7 developmental stages (1 rep each) and 25096 genes
}
\source{
phyexSets::Athaliana.embryogenesis_2011 matched with the Arabidopsis thaliana phylomap from phylomapr
}
\usage{
example_phyex_set_old
}
\description{
Arabidopsis thaliana embryogenesis dataset from Xiang et al. 2011. check the phyexSet package for more details
}
\keyword{datasets}
