% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_phyex_set_sc}
\alias{example_phyex_set_sc}
\title{Example phyex set single cell}
\format{
A ScPhyloExpressionObject with 7 developmental stages (1 rep each) and 25096 genes
}
\source{
generated synthetically, see data-raw
}
\usage{
example_phyex_set_sc
}
\description{
A synthetic single cell for testing
}
\keyword{datasets}
