% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATAI.R
\name{GATAI}
\alias{GATAI}
\title{Low level interface function with GATAI}
\usage{
GATAI(
  ExpressionSet,
  singlecell_data = FALSE,
  gatai_results_path = tempdir(),
  result_overwrite = FALSE,
  condaenv = NULL
)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet, DivergenceExpressionSet or PolymorphismsExpressionSet object.}

\item{singlecell_data}{a logical value indicating whether or not the input \code{ExpressionSet} is a single cell dataset.}

\item{gatai_results_path}{where shall GATAI results be stored. Default \code{gatai_results_path = tempdir()}.}

\item{result_overwrite}{Logical value indicating whether local results of previous GATAI run should be over-written and re-computed?}

\item{condaenv}{The conda environment to use. For \code{\link[reticulate]{use_condaenv}}, this can be the name, the absolute prefix path, or the absolute path to the python binary. If the name is ambiguous, the first environment is used and a warning is issued. For \code{\link[reticulate]{use_miniconda}}, the only conda installation searched is the one installed by \code{\link[reticulate]{install_miniconda}}.}
}
\description{
This function interfaces with the GATAI software which removes the genes that create the \code{\link{PlotSignature}} pattern.
}
\author{
Filipa Martins Costa and Hajk-Georg Drost
}
