% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{get_phylomap}
\alias{get_phylomap}
\title{Extract Phylomap from PhyloExpressionSet}
\usage{
get_phylomap(phyex_set)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}
}
\value{
A tibble with two columns: Stratum (gene age value) and GeneID (gene identifiers)
}
\description{
Extract a phylomap (tibble of strata-to-gene mappings) from a PhyloExpressionSet object.
}
\examples{
# Extract phylomap
phylomap <- extract_phylomap(example_phyex_set)

}
