% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{get_strata_legend}
\alias{get_strata_legend}
\title{Get Strata Legend from PhyloExpressionSet}
\usage{
get_strata_legend(phyex_set)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}
}
\value{
A tibble with two columns: Rank (phylostratum values) and Name (phylostratum labels), sorted by Rank
}
\description{
Extract the strata legend (phylostratum ranks and names) from a PhyloExpressionSet object.
}
\examples{
# Get strata legend
legend <- get_strata_legend(example_phyex_set)

}
