% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test_result.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Goodness of Fit Test}
\usage{
goodness_of_fit(test_result)
}
\arguments{
\item{test_result}{A TestResult object}
}
\value{
A ks.test result object
}
\description{
Perform a Kolmogorov-Smirnov test to assess goodness of fit
between the null sample and fitted distribution.
}
\details{
This function tests whether the null sample follows the fitted distribution
using the Kolmogorov-Smirnov test. A significant result indicates poor fit.
}
\examples{
# Test goodness of fit
# gof_result <- goodness_of_fit(test_result)

}
\keyword{internal}
