% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GroupDiffs.R
\name{GroupDiffs}
\alias{GroupDiffs}
\title{Quantify the significant differences between gene expression distributions of PS or DS groups}
\usage{
GroupDiffs(ExpressionSet, Groups = NULL, legendName = NULL,
  stat.test = "wilcox.test", ...)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{Groups}{a list containing the phylostrata or divergence strata that correspond
to the same phylostratum class or divergence class.
For ex. evolutionary old phylostrata: PS1-3 (Class 1)
and evolutionary young phylostrata: PS4-12 (Class 2). In this case,
the list could be assigned as, \code{Groups} = list(c(1:3), c(4:12)).}

\item{legendName}{a character string specifying whether "PS" or "DS" are used to compute relative expression profiles.}

\item{stat.test}{the statistical test to quantify PS or DS group differences.}

\item{...}{additional plot parameters.}
}
\description{
This function performs a test to quantify the statistical significance between
the global expression level distributions of groups of PS or DS. It therefore, allows users to investigate
significant groups of PS or DS that significantly differ in their gene expression level distibution
within specific developmental stages or experiments.
}
\details{
The purpose of this function is to detect groups of PS or DS that significantly differ in their gene expression
level distributions on a global (transcriptome) level. Since relative expression levels (\code{\link{PlotRE}}) or
PS or DS specific mean expression levels (\code{\link{PlotMeans}}) are biased by highly expressed genes,
this function allows users to objectively test the significant difference of transcriptome expression between
groups of PS or DS in a specific developmental stage or experiment.
}
\examples{
data(PhyloExpressionSetExample)

GroupDiffs(ExpressionSet = PhyloExpressionSetExample,
           Groups       = list(group_1 = 1:3,group_2 = 4:12),
           legendName   = "PS")
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{PlotGroupDiffs}}, \code{\link{PlotMeans}}, \code{\link{PlotRE}}, \code{\link{PlotBarRE}}, \code{\link{PlotCategoryExpr}}
}

