% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harm.mean.R
\name{harm.mean}
\alias{harm.mean}
\title{Harmonic Mean}
\usage{
harm.mean(x)
}
\arguments{
\item{x}{a numeric vector for which harmonic mean computations shall be performed.}
}
\description{
This function computes the harmonic mean of a numeric input vector \code{x}.
}
\examples{
x <- 1:10

harm.mean(x)

}
\author{
Hajk-Georg Drost
}
