% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.ExpressionSet}
\alias{is.ExpressionSet}
\title{Function testing for the validaty of a PhyloExpressionSet or DivergenceExpressionSet standard}
\usage{
is.ExpressionSet(ExpressionSet)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet that shall be tested for format validity.}
}
\description{
This function tests whether a given ExpressionSet follows the pre-defined PhyloExpressionSet or DivergenceExpressionSet standard.
}
\examples{
# read example PhyloExpressionSet
data(PhyloExpressionSetExample)

is.ExpressionSet(PhyloExpressionSetExample)
}
\author{
Hajk-Georg Drost
}

