% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed_gatai.R
\name{is_installed_gatai}
\alias{is_installed_gatai}
\title{Test whether GATAI is installed on user machine}
\usage{
is_installed_gatai()
}
\description{
Helper function to test if the user has set up GATAI correctly in their environment.
}
