% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_example_phyex_set_sc}
\alias{load_example_phyex_set_sc}
\title{Load Example Single-Cell PhyloExpressionSet}
\usage{
load_example_phyex_set_sc()
}
\value{
A ScPhyloExpressionSet object, or NULL if Seurat is not available.
}
\description{
Creates and returns an example ScPhyloExpressionSet object for testing and examples.
Only works if Seurat is installed.
}
