% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_utils.R
\name{new_required_property}
\alias{new_required_property}
\title{Create S7 Required Property}
\usage{
new_required_property(class = S7::class_any, name, validator = NULL)
}
\arguments{
\item{class}{The S7 class for the property (default: S7::class_any)}

\item{name}{Name of the property (used in error messages)}

\item{validator}{Optional validation function for the property}
}
\value{
An S7 property object that is required
}
\description{
Helper function to create an S7 property that is required (throws error if not provided).
}
\examples{
# Create a required property
# required_prop <- new_required_property(class = class_character, name = "gene_ids")

}
\keyword{internal}
