% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_transforms.R
\name{normalise_stage_expression}
\alias{normalise_stage_expression}
\title{Normalise Stage Expression Data}
\usage{
normalise_stage_expression(phyex_set, total = 1e+06)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{total}{Numeric value to normalise each sample to (default: 1e6)}
}
\value{
A PhyloExpressionSet object with normalised expression data
}
\description{
Normalise expression data to a specified total expression level per sample.
}
\examples{
# Normalise to 1 million total expression per sample
normalised_set <- normalise_stage_expression(example_phyex_set, total = 1e6)

}
