% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseTest.R
\name{PairwiseTest}
\alias{PairwiseTest}
\title{Perform Pairwise Difference Test}
\usage{
PairwiseTest(
  ExpressionSet,
  modules = NULL,
  altHypothesis = NULL,
  permutations = 1000,
  lillie.test = FALSE,
  plotHistogram = FALSE,
  runs = 10,
  parallel = FALSE,
  gof.warning = FALSE,
  custom.perm.matrix = NULL
)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{modules}{a list storing two elements: contrast1 and contrast2. Each element expects a numeric
vector specifying the developmental stages or experiments that correspond to each module. 
For example, \code{module} = list(contrast1 = 1:2, contrast2 = 3:7) divides a dataset 
storing seven developmental stages into 2 modules (contrasts).}

\item{altHypothesis}{a character string defining the alternative hypothesis (i.e. direction of difference)
used to quantify the statistical significance in the present phylotranscriptomics pattern.
Possible values can be:
\itemize{
\item \code{altHypothesis} = \code{"greater"} : contrast1 > contrast2
\item \code{altHypothesis} = \code{"less"} : contrast1 < contrast2
}}

\item{permutations}{a numeric value specifying the number of permutations to be performed for the \code{ReductiveHourglassTest}.}

\item{lillie.test}{a boolean value specifying whether the Lilliefors Kolmogorov-Smirnov Test shall be performed to quantify the goodness of fit.}

\item{plotHistogram}{a boolean value specifying whether a \emph{Lillifor's Kolmogorov-Smirnov-Test} 
shall be performed to test the goodness of fit of the approximated distribution, as well as additional plots quantifying the significance
of the observed phylotranscriptomic pattern.}

\item{runs}{specify the number of runs to be performed for goodness of fit computations, in case \code{plotHistogram} = \code{TRUE}.
In most cases \code{runs} = 100 is a reasonable choice. Default is \code{runs} = 10 (because it takes less computation time for demonstration purposes).}

\item{parallel}{performing \code{runs} in parallel (takes all cores of your multicore machine).}

\item{gof.warning}{a logical value indicating whether non significant goodness of fit results should be printed as warning. Default is \code{gof.warning = FALSE}.}

\item{custom.perm.matrix}{a custom \code{\link{bootMatrix}} (permutation matrix) to perform the underlying test statistic. Default is \code{custom.perm.matrix = NULL}.}
}
\value{
a list object containing the list elements:

\code{p.value} : the p-value quantifying the statistical significance of the given phylotranscriptomics pattern.

\code{std.dev} : the standard deviation of the N sampled phylotranscriptomics patterns for each developmental stage S.

\code{lillie.test} : a boolean value specifying whether the \emph{Lillifors KS-Test} returned a p-value > 0.05, 
which indicates that fitting the permuted scores with a normal distribution seems plausible.
}
\description{
The \emph{Pairwise Difference Test} aims to statistically evaluate the pairwise
difference in the phylotranscriptomic pattern between two contrasts based on \code{\link{TAI}} or \code{\link{TDI}} computations.
The corresponding p-value quantifies the probability that a given TAI or TDI pattern (or any phylotranscriptomics pattern) 
does not differ from the alternative hypothesis (specifying the direction of difference).
A p-value < 0.05 indicates that the corresponding phylotranscriptomics pattern does
indeed differ.
}
\details{
The \emph{reductive pairwise difference test} is a permutation test based on the following test statistic. 

(1) A PhyloExpressionSet is partitioned into contrast pairs - contrast1 and contrast2 - based on prior biological knowledge. 
This prior knowledge could include sexual, ecological and genetic backgrounds.

(2) The mean \code{\link{TAI}} or \code{\link{TDI}} value for each of the two contrasts contrast1 and contrast2 are computed. 

(3) The pairwise differences D_contrast = contrast1 - contrast2 is calculated as final test statistic of the pairwise test,
when \code{altHypothesis} is specified as "greater". When \code{altHypothesis} is specified as "less", sign of D_contrast is reversed.


In order to determine the statistical significance of an observed pairwise difference D_contrast 
the following permutation test was performed. Based on the \code{\link{bootMatrix}} D_contrast 
is calculated from each of the permuted \code{\link{TAI}} or \code{\link{TDI}} profiles, 
approximated by a Gaussian distribution with method of moments estimated parameters returned by \code{\link[fitdistrplus]{fitdist}}, 
and the corresponding p-value is computed by \code{\link{pnorm}} given the estimated parameters of the Gaussian distribution. 
The \emph{goodness of fit} for the random vector \emph{D_contrast} is statistically quantified by an Lilliefors (Kolmogorov-Smirnov) test 
for normality.


In case the parameter \emph{plotHistogram = TRUE}, a multi-plot is generated showing:
       
(1) A Cullen and Frey skewness-kurtosis plot generated by \code{\link[fitdistrplus]{descdist}}. 
This plot illustrates which distributions seem plausible to fit the resulting permutation vector D_contrast 
In the case of the \emph{pairwise difference test} a normal distribution seemed plausible.

(2) A histogram of D_contrast combined with the density plot is plotted. D_contrast is then fitted by a normal distribution. 
The corresponding parameters are estimated by \emph{moment matching estimation} using the \code{\link[fitdistrplus]{fitdist}} function.

(3) A plot showing the p-values for N independent runs to verify that a specific p-value is biased by a specific permutation order.

(4) A barplot showing the number of cases in which the underlying goodness of fit (returned by Lilliefors (Kolmogorov-Smirnov) test 
for normality) has shown to be significant (\code{TRUE}) or not significant (\code{FALSE}). 
This allows to quantify the permutation bias and their implications on the goodness of fit.
}
\examples{

data(PhyloExpressionSetExample)

# perform the pairwise difference test for a PhyloExpressionSet
# here the prior biological knowledge is that stages 1-2 correspond to contrast1,
# stages 3-7 correspond to contrast 2.
# We test whether TAI in contrast1 is greater than contrast 2.
PairwiseTest(PhyloExpressionSetExample,
         modules = list(contrast1 = 1:2, contrast2 = 3:7), 
         altHypothesis = "greater",
         permutations = 1000)

# We can also test whether TAI in contrast1 is less than contrast 2.
PairwiseTest(PhyloExpressionSetExample,
         modules = list(contrast1 = 1:2, contrast2 = 3:7), 
         altHypothesis = "less",
         permutations = 1000)

# if we only want to test whether TAI in stage 1 (contrast 1) is greater than stage 3 (contrast 2).
PairwiseTest(PhyloExpressionSetExample,
         modules = list(contrast1 = 1, contrast2 = 2), 
         altHypothesis = "greater",
         permutations = 1000)

# use your own permutation matrix based on which p-values (PairwiseTest)
# shall be computed
custom_perm_matrix <- bootMatrix(PhyloExpressionSetExample,100)
# We test whether TAI in contrast1 is greater than contrast 2.
PairwiseTest(PhyloExpressionSetExample,
         modules = list(contrast1 = 1:2, contrast2 = 3:7), 
         altHypothesis = "greater",
         custom.perm.matrix = custom_perm_matrix)
                       
}
\references{
Drost HG et al. (2015) Mol Biol Evol. 32 (5): 1221-1231 doi:10.1093/molbev/msv012

Lotharukpong JS et al. (2024) (unpublished)
}
\seealso{
\code{\link{lcScore}}, \code{\link{bootMatrix}}, \code{\link{FlatLineTest}},\code{\link{ReductiveHourglassTest}}, \code{\link{ReverseHourglassTest}}, \code{\link{PlotSignature}}
}
\author{
Hajk-Georg Drost and Jaruwatana Sodai Lotharukpong
}
