% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{permute_PS}
\alias{permute_PS}
\title{Permute Strata in PhyloExpressionSet}
\usage{
permute_PS(phyex_set)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}
}
\value{
A new PhyloExpressionSet object with permuted strata and strata_values
}
\description{
Returns a copy of the PhyloExpressionSet with permuted strata and corresponding strata values.
}
