% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{PhyloExpressionSetBase}
\alias{PhyloExpressionSetBase}
\title{PhyloExpressionSet Base Class}
\usage{
PhyloExpressionSetBase(
  strata = stop("@strata is required"),
  gene_ids = stop("@gene_ids is required"),
  name = "Phylo Expression Set",
  species = character(0),
  index_type = "TXI",
  identities_label = "Identities",
  null_conservation_sample_size = 5000L,
  precomputed_null_conservation_txis = NULL
)
}
\description{
Abstract S7 base class for storing and manipulating phylotranscriptomic expression data.
This class provides the common interface for both bulk and single-cell phylotranscriptomic data.
}
