% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylostratigraphy.R
\name{Phylostratigraphy}
\alias{Phylostratigraphy}
\title{Perform Phylostratigraphy}
\usage{
Phylostratigraphy(query, sci.name, seq.type = "protein",
  download.databases = FALSE, max.target.seqs = 1e+06, eval = "1E-5",
  cores = 1, custom.taxa = NULL, custom.db = NULL, nr.path = NULL,
  tax.db.path = NULL, add.blast.parameters = NULL,
  add.makeblastdb.params = NULL, blast.path = NULL)
}
\arguments{
\item{query}{a character string specifying the path to the protein file of interest (query organism).}

\item{sci.name}{the scientific name of the query organims.}

\item{seq.type}{the sequence type of the input file (usually 'protein') for Phylostratigraphy.}

\item{download.databases}{either a logical in case both \code{NCBI nr} and \code{NCBI Taxonomy} should
be loaded automatically to your local machine or a character string specifying which of the both databases
shall be downloaded, \code{download.databases = "nr"} (\code{NCBI nr}) or \code{download.databases = "taxdb"} (\code{NCBI Taxonomy})}

\item{max.target.seqs}{a numeric value specifying the number of aligned sequences to keep.}

\item{eval}{E-value cutoff for blastp searches. Default is \code{eval = "1E-5"}.}

\item{cores}{cores for parallel computations.}

\item{custom.taxa}{a character vector storing all taxon names for which tax_ids shall be retrieved and gene age assignment based on.}

\item{custom.db}{a character string specifying the path to the folder in which the customized
BLAST database (in FASTA file format) is stored. This customized database is then formatted by \code{makeblastdb} and Phylostratigraphy is performed against this formatted (custom) database.}

\item{nr.path}{path to the NCBI nr database folder. This can also be specified when \code{download.databases = TRUE} to specify the folder in which the database shall be stored.}

\item{tax.db.path}{path to the NCBI Taxonomy database folder. This can also be specified when \code{download.databases = TRUE} to specify the folder in which the database shall be stored.}

\item{add.blast.parameters}{a character string listing the input paramters that shall be passed to the executing BLAST program. Default is \code{NULL}, implicating that a set of default parameters is used when running BLAST.}

\item{add.makeblastdb.params}{a character string listing the input paramters that shall be passed to the executing makeblastdb program. Default is \code{NULL}, implicating that a set of default parameters is used when running makeblastdb}

\item{blast.path}{a character string specifying the path to the BLAST program (in case users don't use the default blast path).}
}
\description{
This function takes the protein sequence file of a query organism 
and performs the \code{Phylostratigraphy} algorithm to the \code{NCBI nr} database.
}
\details{
This
}
\examples{
\dontrun{
# Construct a phylostratigraphic map for 'Arabidopsis thaliana'
Athaliana.PhyloMap <- Phylostratigraphy(
   query              = system.file('seqs/example_athal_aa.faa', package = 'orthologr'),
   sci.name           = "Arabidopsis thaliana",
   seq.type           = "protein", 
   download.databases = FALSE, 
   max.target.seqs    = 1000000, 
   eval               = "1E-5", 
   cores              = 1, 
   nr.path            = "path/to/local/nr/database/",
   tax.db.path        = "path/to/local/Taxonomy/database/")

}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{DivergenceStratigraphy}}, \code{\link{TAI}}, \code{\link{RE}}, \code{\link{PlotRE}},
\code{\link{PlotPattern}}
}

