% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_partialTAI.R
\name{plot_distribution_partialTAI}
\alias{plot_distribution_partialTAI}
\title{Comparing partial TAI distributions across the different developmental stages}
\usage{
plot_distribution_partialTAI(
  ExpressionSet,
  stages = 1:ncol(ExpressionSet) - 2,
  xlab = "Partial TAI",
  ylab = "Density",
  main = "Density Distribution of Partial TAI by Developmental Stage",
  seed = 123
)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{stages}{a numeric vector specifying the indices of the stages to compare. Each index 
corresponds to a stage in the ExpressionSet. Starts in one.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{figure title.}

\item{seed}{defines the colors for the different developmetal stages}
}
\description{
\emph{plot_distribution_partialTAI} generates 2 plots that help to compare the distribution
of partial TAI through various developmental stages, highlighting each stage with 
distinct colors.
}
\author{
Filipa Martins Costa
}
