% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_partialTAI_qqplot.R
\name{plot_distribution_partialTAI_qqplot}
\alias{plot_distribution_partialTAI_qqplot}
\title{QQ plot comparing partial TAI distributions across developmental stages against a reference stage}
\usage{
plot_distribution_partialTAI_qqplot(
  ExpressionSet,
  reference_stage_index = 1,
  xlab = "Quantiles of Reference Stage",
  ylab = "Quantiles of Other Stages",
  main = "QQ Plot: Developmental Stages vs Reference Stage (p-values from KS test)",
  alpha = 0.7,
  size = 1.2
)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{reference_stage_index}{an integer specifying the index of the reference developmental stage 
of the ExpressionSet. The partial TAI distribution of this stage will be used as the reference 
for comparisons with other stages (as default stage index 1).}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{figure title.}

\item{alpha}{transparency of the point.}

\item{size}{size of the points.}
}
\description{
\emph{plot_distribution_partialTAI_qqplot} generates a QQ plot to compare the partial TAI 
distributions of various developmental stages against a reference stage.
It visualizes quantile differences between the reference and other stages,
highlights each stage with distinct colors, and annotates the plot with the p-values
from the nonparametric \code{\link{ks.test}} to indicate the significance of distribution differences.
}
\author{
Filipa Martins Costa
}
