% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_multiple.R
\name{plot_signature_multiple}
\alias{plot_signature_multiple}
\title{Plot Multiple Transcriptomic Signatures}
\usage{
plot_signature_multiple(
  phyex_sets,
  legend_title = "Phylo Expression Set",
  show_p_val = F,
  conservation_test = flatline_test,
  transformation = NULL,
  colours = NULL,
  ...
)
}
\arguments{
\item{phyex_sets}{A list of PhyloExpressionSet objects to compare}

\item{legend_title}{Title for the legend (default: "Phylo Expression Set")}

\item{show_p_val}{Logical indicating whether to show p-values (default: FALSE)}

\item{conservation_test}{Function to use for conservation testing (default: flatline_test)}

\item{transformation}{Optional transformation function to apply to all datasets (default: NULL)}

\item{colours}{Optional vector of colors for each dataset (default: NULL)}

\item{...}{Additional arguments passed to plot_signature}
}
\value{
A ggplot2 object showing multiple transcriptomic signatures
}
\description{
Create a plot comparing multiple transcriptomic signatures on the same axes,
with options for statistical testing and transformations.
}
\details{
This function allows comparison of multiple transcriptomic signatures by overlaying
them on the same plot. Each signature is colored differently and can be tested
for conservation patterns. If a transformation is provided, it's applied to all
datasets before plotting.
}
\examples{
# Compare multiple datasets
# phyex_list <- list(dataset1 = phyex_set1, dataset2 = phyex_set2)
# p <- plot_signature_multiple(phyex_list, legend_title = "Dataset")

# With transformation
# p2 <- plot_signature_multiple(phyex_list, transformation = log1p)


}
