% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotCategoryExprBars.R
\name{PlotCategoryExprBars}
\alias{PlotCategoryExprBars}
\title{Plot the Expression Levels of each Age or Divergence Category as Barplot or Violinplot}
\usage{
PlotCategoryExprBars(ExpressionSet, legendName, type = "category-centered",
  distr.type = "barplot", log.expr = FALSE)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{legendName}{a character string specifying whether "PS" or "DS" are used to compute relative expression profiles.}

\item{type}{type of age or divergence category comparison. Specifications can be \code{type = "category-centered"} or \code{type = "stage-centered"}.}

\item{distr.type}{format of visualizing age or divergence category specific expression distributions. Either \code{distr.type = "barplot"} or
\code{distr.type = "violin"}.}

\item{log.expr}{a logical value specifying whether or not expression levels should be log2-transformed before visualization.}
}
\description{
This function visualizes the expression level distribution of each phylostratum during each time point or experiment
as barplot or violin plot enabling users to quantify the age (PS) or divergence category (DS) specific contribution to the
corresponding transcriptome.
}
\details{
This way of visualizing the gene expression distribution of each age (PS) or divergence category (DS) during
all developmental stages or experiments allows users to detect specific age or divergence categories contributing significant
levels of gene expression to the underlying biological process (transcriptome).

This quantification allows to conclude that genes originating in specific PS or DS contribute significantly more to the overall transcriptome
than other genes originating from different PS or DS categories. More specialized analyses such as \code{\link{PlotMeans}}, \code{\link{PlotRE}},
\code{\link{PlotBarRE}}, etc. will then allow to study the exact mean expression patterns of these age or divergence categories.

Argument Specifications:

Argument: type

\itemize{
\item \code{type = "category-centered"} This specification allows users to compare the differences between all age or
 divergence categories during the same stage or experiment.
 \item \code{type = "stage-centered"} This specification allows users to compare the differences between all age or
 divergence categories between stages or experiments.
}


Argument: distr.type

\itemize{
\item \code{distr.type = "barplot"} This specification allows users to visualize the expression distribution of all PS or DS as barplot.
\item \code{distr.type = "violin"} This specification allows users to visualize the expression distribution of all PS or DS as violin plot.
}
}
\examples{
data(PhyloExpressionSetExample)
data(DivergenceExpressionSetExample)

# category-centered visualization of PS specific expression level distributions (log-scale)
PlotCategoryExprBars(ExpressionSet = PhyloExpressionSetExample,
                     legendName    = "PS",
                     type          = "category-centered",
                     distr.type    = "barplot",
                     log.expr      = TRUE)



# stage-centered visualization of PS specific expression level distributions (log-scale)
PlotCategoryExprBars(ExpressionSet = PhyloExpressionSetExample,
                     legendName    = "PS",
                     distr.type    = "barplot",
                     type          = "stage-centered",
                     log.expr      = TRUE)



# category-centered visualization of PS specific expression level distributions (log-scale)
# as violoin plot
PlotCategoryExprBars(ExpressionSet = PhyloExpressionSetExample,
                     legendName    = "PS",
                     distr.type    = "violin",
                     type          = "stage-centered",
                     log.expr      = TRUE)




# analogous for DivergenceExpressionSets
PlotCategoryExprBars(ExpressionSet = DivergenceExpressionSetExample,
                     legendName    = "DS",
                     type          = "category-centered",
                     distr.type    = "barplot",
                     log.expr      = TRUE)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{PlotMeans}}, \code{\link{PlotRE}}, \code{\link{PlotBarRE}}, \code{\link{age.apply}}
}

