% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCIRatio.R
\name{PlotCIRatio}
\alias{PlotCIRatio}
\title{Plot Transcriptome Index using bootstrapping and confidence intervals}
\usage{
PlotCIRatio(ExpressionSet, measure, nbootstraps)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet, DivergenceExpressionSet or PolymorphismsExpressionSet object.}

\item{measure}{type of transcriptome index that shall be computed. E.g. measure = "TAI" (Transcriptome Age Index), measure = "TDI" (Transcriptome Divergence Index), measure = "TPI" (Transcriptome Polymorphism Index).}

\item{nbootstraps}{number of independent bootstraps.}
}
\description{
Function to plot Transcriptome Index using bootstrapping and confidence intervals instead of permutation tests used in \code{\link{PlotSignature}}.
}
\details{
The
}
\examples{
data("PhyloExpressionSetExample")
PlotCIRatio(PhyloExpressionSetExample, measure = "TAI", nbootstraps)
}
\seealso{
\code{\link{PlotSignature}}
}
\author{
Jialin Liu
}
