% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotEnrichment.R
\name{PlotEnrichment}
\alias{PlotEnrichment}
\title{Plot the Phylostratum or Divergence Stratum Enrichment of a given Gene Set}
\usage{
PlotEnrichment(ExpressionSet, test.set, measure = "log-foldchange",
  legendName = NULL, over.col = "steelblue", under.col = "midnightblue",
  epsilon = 1e-05, cex.legend = 1, cex.asterisk = 1, plot.bars = TRUE,
  ...)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{test.set}{a character vector storing the gene ids for which PS/DS enrichment analyses should be performed.}

\item{measure}{a character string specifying the measure that should be used to quantify over and under representation of PS/DS.}

\item{legendName}{a character string specifying whether "PS" or "DS" are used to compute relative expression profiles.}

\item{over.col}{color of the overrepresentation bars.}

\item{under.col}{color of the underrepresentation bars.}

\item{epsilon}{a small value to shift values by epsilon to avoid log(0) computations.}

\item{cex.legend}{the \code{cex} value for the legend.}

\item{cex.asterisk}{the \code{cex} value for the asterisk.}

\item{plot.bars}{a logical value specifying whether or not bars should be visualized or whether only \code{p.values} and \code{enrichment.matrix} should be returned.}

\item{...}{default graphics parameters.}
}
\description{
This function computes and visualizes the significance of enriched (over or underrepresented) Phylostrata or Divergence Strata within an input \code{test.set}.
}
\details{
This
}
\examples{
data(PhyloExpressionSetExample)

set.seed(123)
test_set <- sample(PhyloExpressionSetExample[ , 2],1000)

# measure: log-foldchange
PlotEnrichment(ExpressionSet = PhyloExpressionSetExample,
               test.set      = test_set ,
               legendName    = "PS",
               measure       = "log-foldchange")


# measure: foldchange
PlotEnrichment(ExpressionSet = PhyloExpressionSetExample,
               test.set      = test_set ,
               legendName    = "PS",
               measure       = "foldchange")
}
\author{
Hajk-Georg Drost
}

