% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotGeneSet.R
\name{PlotGeneSet}
\alias{PlotGeneSet}
\title{Plot the Expression Profiles of a Gene Set}
\usage{
PlotGeneSet(ExpressionSet, gene.set, get.subset = FALSE, colors = NULL, ...)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{gene.set}{a character vector storing the gene ids for which gene expression profiles shall be visualized.}

\item{get.subset}{a logical value indicating whether or not an \code{ExpressionSet} subset of the selected \code{gene.set} should be retuned.}

\item{colors}{colors for gene expression profiles. Default: \code{colors = NULL}, hence default colours are used.}

\item{...}{additional parameters passed to \code{\link{matplot}}.}
}
\description{
This function simply visualizes the gene expression profiles of
a defined subset of genes stored in the input \code{ExpressionSet}.
}
\examples{
data(PhyloExpressionSetExample)

# png("test_png.png",700,400)
PlotGeneSet(ExpressionSet = PhyloExpressionSetExample,
            gene.set      = PhyloExpressionSetExample[1:5, 2],
            type          = "l",
            lty           = 1,
            lwd           = 4,
            xlab          = "Ontogeny",
            ylab          = "Expression Level")

# dev.off()
}
\author{
Hajk-Georg Drost
}

