% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotReplicateQuality.R
\name{PlotReplicateQuality}
\alias{PlotReplicateQuality}
\title{Plot the Quality of Biological Replicates}
\usage{
PlotReplicateQuality(ExpressionSet, nrep, stage.names, ...)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{nrep}{either a numeric value specifying the constant number of replicates per stage or a numeric vector specifying the variable number of replicates for each stage position.}

\item{stage.names}{a character vector specifying the new names of collapsed stages.}

\item{...}{additional graphics parameters.}
}
\description{
This function performs several quality checks to validate the
biological variation between replicates and stages (experiments).
}
\details{
The following quality checks can be performed:
\itemize{
\item
\item
}
}
\author{
Hajk-Georg Drost
}

