% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{PS_colours}
\alias{PS_colours}
\title{Generate Phylostratum Colors}
\usage{
PS_colours(n)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
A character vector of color codes
}
\description{
Generate a color palette for phylostrata visualization using a log-scaled transformation.
}
\examples{
# Generate colors for 5 phylostrata
colors <- PS_colours(5)
}
