% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStrataTEI.R
\name{pStrataTEI}
\alias{pStrataTEI}
\title{Compute Partial Transcriptome Evolutionary Index (TEI) Strata Values}
\usage{
pStrataTEI(
  ExpressionSet,
  Phylostratum = NULL,
  split = 1e+05,
  showprogress = TRUE,
  threads = 1
)
}
\arguments{
\item{ExpressionSet}{expression object with rownames as GeneID (dgCMatrix)
or standard PhyloExpressionSet object.}

\item{Phylostratum}{a named vector representing phylostratum per GeneID with
names as GeneID (not used if Expression is PhyloExpressionSet).}

\item{split}{specify number of columns to split}

\item{showprogress}{boolean if progressbar should be shown}

\item{threads}{specify number of threads}
}
\value{
a numeric matrix storing the summed partial TEI values for each
strata.
}
\description{
This function computes the partial transcriptome evolutionary
index (TEI) values combined for each strata.

In detail, each gene gets a \emph{TEI contribution profile}.

\deqn{TEI_is = f_is * ps_i}

where TEI_is is the partial TEI value of gene i,
\eqn{f_is = e_is / \sum e_is} and \eqn{ps_i} is the phylostratum of gene i.

\eqn{TEI_is} values are combined per \eqn{ps}.
}
\details{
The partial TEI values combined per strata give an overall
impression of the contribution of each
strata to the global \code{\link{TEI}} pattern.
}
\examples{

# reading a standard PhyloExpressionSet
data(PhyloExpressionSetExample, package = "myTAI")

# computing partial TEI contribution per gene
pS <- pStrataTEI(PhyloExpressionSetExample)

}
\references{
Domazet-Loso T. and Tautz D. (2010).
\emph{A phylogenetically based transcriptome age index mirrors ontogenetic
divergence patterns}. Nature (468): 815-818.

Quint M et al. (2012).
\emph{A transcriptomic hourglass in plant embryogenesis}.
Nature (490): 98-101.

Drost HG et al. (2015)
Mol Biol Evol. 32 (5): 1221-1231 doi:10.1093/molbev/msv012
}
\author{
Kristian K Ullrich
}
