% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_boottei_parallel}
\alias{rcpp_boottei_parallel}
\title{rcpp_boottei_parallel}
\usage{
rcpp_boottei_parallel(expression, ps, permutations, ncores = 1L)
}
\arguments{
\item{expression}{ExpressionSet as sparseMatrix}

\item{ps}{named Phylostratum}

\item{permutations}{number of permutations}

\item{ncores}{number of cores}
}
\value{
sparseMatrix
}
\description{
computes the phylogenetically based
transcriptome evolutionary index (TEI) shuffling the strata for permutation
statistic
}
\examples{
## load example PhyloExpressionSetExample

data("PhyloExpressionSetExample", package="myTAI")

## convert into sparseMatrix - rownames GeneID

spmat <- as(data.matrix(PhyloExpressionSetExample[,-c(1,2)]),
    "sparseMatrix")
rownames(spmat) <- PhyloExpressionSetExample$GeneID

## create named Phylostratum vector

ps <- setNames(PhyloExpressionSetExample$Phylostratum,
    PhyloExpressionSetExample$GeneID)

## get permutations
rcpp_boottei_parallel(spmat, ps, 100, 1)
}
\author{
Kristian K Ullrich
}
