% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_expression.R
\name{relative_expression}
\alias{relative_expression}
\title{Relative Expression Functions}
\usage{
relative_expression(count_matrix)
}
\arguments{
\item{count_matrix}{A numeric matrix where columns represent developmental stages and rows represent genes.}
}
\value{
A numeric vector of relative expression values for each stage (column) in the input matrix.
}
\description{
Functions for computing and plotting relative expression profiles using PhyloExpressionSet S7 objects.

Computes the relative expression profile for a given gene expression matrix. 
The relative expression is calculated by normalizing the column means of the matrix to a [0, 1] scale.
}
