% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{rename_phyex_set}
\alias{rename_phyex_set}
\title{Rename a PhyloExpressionSet}
\usage{
rename_phyex_set(phyex_set, new_name)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{new_name}{Character string for the new dataset name}
}
\value{
The PhyloExpressionSet object with the updated name
}
\description{
Returns a copy of the PhyloExpressionSet with a new name.
}
