% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{reorder_identities}
\alias{reorder_identities}
\title{Set Identity Order for ScPhyloExpressionSet}
\usage{
reorder_identities(phyex_set, new_order)
}
\arguments{
\item{phyex_set}{A ScPhyloExpressionSet object}

\item{new_order}{Character vector specifying the new order of identities}
}
\value{
ScPhyloExpressionSet object with updated identity order
}
\description{
Change the order of the identities (factor levels) in the Seurat object.
}
