% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{select_genes}
\alias{select_genes}
\title{Select Genes from PhyloExpressionSet}
\usage{
select_genes(phyex_set, ...)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{...}{Additional arguments passed to methods (typically includes 'genes' parameter)}
}
\value{
A PhyloExpressionSet object containing only the selected genes
}
\description{
Extract a subset of genes from a PhyloExpressionSet object.
}
\examples{
# Select specific genes
selected_set <- select_genes(example_phyex_set, example_phyex_set@gene_ids[1:10])

}
