% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_transforms.R
\name{set_expression}
\alias{set_expression}
\title{Gene Expression Transformation Functions}
\usage{
set_expression(phyex_set, ...)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{...}{Additional arguments}

\item{new_expression}{Matrix to set as the new expression}
}
\value{
A PhyloExpressionSet object with updated expression
}
\description{
Collection of functions for transforming gene expression data
in PhyloExpressionSet objects.

Generic function to set the expression matrix in a PhyloExpressionSet object.
}
