% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{set_identities}
\alias{set_identities}
\title{Set Identities for ScPhyloExpressionSet}
\usage{
set_identities(phyex_set, identity_name)
}
\arguments{
\item{phyex_set}{A ScPhyloExpressionSet object}

\item{identity_name}{Character, name of the metadata column to use as new identities}
}
\value{
ScPhyloExpressionSet object with updated identities
}
\description{
Change the Seurat identities to a different metadata column.
}
