% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{set_identity_colours}
\alias{set_identity_colours}
\title{Set Identity Colours for ScPhyloExpressionSet}
\usage{
set_identity_colours(phyex_set, identity_name, colours)
}
\arguments{
\item{phyex_set}{A ScPhyloExpressionSet object}

\item{identity_name}{Character, name of the metadata column to set colours for}

\item{colours}{Named character vector of colours where names correspond to identity values}
}
\value{
ScPhyloExpressionSet object with updated identity colours
}
\description{
Set custom colours for one or more identity columns in the ScPhyloExpressionSet object.
}
\examples{
# Set colours for specific identity
# colours <- c("TypeA" = "red", "TypeB" = "blue", "TypeC" = "green")
# sc_set <- set_identity_colours(sc_set, "groups", colours)
}
