% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_null_conservation_txis.R
\name{stat_generate_conservation_txis}
\alias{stat_generate_conservation_txis}
\title{Generate Null Conservation TXI Distribution}
\usage{
stat_generate_conservation_txis(strata_vector, count_matrix, sample_size)
}
\arguments{
\item{strata_vector}{Numeric vector of phylostratum assignments}

\item{count_matrix}{Matrix of expression counts}

\item{sample_size}{Number of permutations to generate}
}
\value{
Matrix of permuted TXI values for null hypothesis testing
}
\description{
Generate a null distribution of transcriptomic index values for
conservation testing by permuting phylostratum assignments.
}
\details{
This function creates a null distribution by randomly permuting phylostratum
assignments while keeping the expression matrix fixed. This preserves the
expression structure while breaking the phylostratum-expression relationship.
# Generate null TXI distribution
# null_txis <- stat_generate_conservation_txis(strata_vec, expr_matrix, 1000)
}
\keyword{internal}
