% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{TEI}
\alias{TEI}
\title{Calculate Transcriptomic Evolutionary Index (TEI)}
\usage{
TEI(phyex_set)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}
}
\value{
Numeric vector of TEI values for each condition
}
\description{
Calculate the transcriptomic evolutionary index values for a PhyloExpressionSet.
This function provides backward compatibility with the old TEI() function.
}
\examples{
# Calculate TEI values
# tei_values <- TEI(phyex_set)

}
