% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfStability.R
\name{tfStability}
\alias{tfStability}
\title{Perform Permutation Tests Under Different Transformations}
\usage{
tfStability(
  ExpressionSet,
  TestStatistic = "FlatLineTest",
  transforms = c("log1p", "sqrt", "none"),
  modules = NULL,
  permutations = 1000,
  pseudocount = 0
)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{transforms}{a character vector of any valid function that transforms gene expression levels.}

\item{modules}{a list storing three elements: early, mid, and late. Each element expects a numeric
vector specifying the developmental stages or experiments that correspond to each module. 
For example, \code{module} = list(early = 1:2, mid = 3:5, late = 6:7) devides a dataset 
storing seven developmental stages into 3 modules.}

\item{permutations}{a numeric value specifying the number of permutations to be performed for the \code{ReductiveHourglassTest}.}

\item{pseudocount}{any valid number to the expression matrix prior to transformation.}
}
\value{
a vector object containing the vector elements:

\code{p.value} : the p-value quantifying the statistical significance (depending on the chosen test) of the given phylotranscriptomics pattern under the given data transformation(s).
}
\description{
\emph{tfStability} aims to statistically evaluate the
stability of \code{\link{ReductiveHourglassTest}}, \code{\link{FlatLineTest}}, 
\code{\link{ReverseHourglassTest}}, or \code{\link{EarlyConservationTest}}
(all based on \code{\link{TAI}} or \code{\link{TDI}} computations) against different
data transformations.
The corresponding p-value quantifies the probability that a given TAI or TDI pattern (or any phylotranscriptomics pattern) 
does not support the chosen test. A p-value < 0.05 indicates that the corresponding phylotranscriptomics pattern does
indeed support the chosen test.
}
\details{
An assessment for the stability of a data transforms on the results of the permutation test of choice.
The reductive hourglass test is a permutation test based on the following test statistic. 
For details, please consult \code{\link{ReductiveHourglassTest}}, \code{\link{FlatLineTest}}, 
\code{\link{ReverseHourglassTest}} or \code{\link{EarlyConservationTest}}
}
\examples{

data(PhyloExpressionSetExample)

# perform the reductive hourglass test for a PhyloExpressionSet
# here the prior biological knowledge is that stages 1-2 correspond to module 1 = early,
# stages 3-5 to module 2 = mid (phylotypic module), and stages 6-7 correspond to
# module 3 = late
tfStability(ExpressionSet = PhyloExpressionSetExample,
                     TestStatistic = "ReverseHourglassTest",
                     transforms = c("log1p", "sqrt", "none"),
                     modules = list(early = 1:2, mid = 3:5, late = 6:7))


}
\references{
Lotharukpong JS et al. (2022) bioRxiv
}
\seealso{
\code{\link{rhScore}}, \code{\link{bootMatrix}}, \code{\link{FlatLineTest}},
\code{\link{ReverseHourglassTest}}, \code{\link{EarlyConservationTest}}, 
\code{\link{ReductiveHourglassTest}}, \code{\link{PlotSignature}},
}
\author{
Jaruwatana Sodai Lotharukpong
}
