% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\docType{data}
\name{TI_map}
\alias{TI_map}
\title{Transcriptomic Index Name Mapping}
\format{
A named list with 5 elements:
\describe{
  \item{TXI}{Transcriptomic Index}
  \item{TAI}{Transcriptomic Age Index}
  \item{TDI}{Transcriptomic Divergence Index}
  \item{TPI}{Transcriptomic Polymorphism Index}
  \item{TEI}{Transcriptomic Evolutionary Index}
}
}
\usage{
TI_map
}
\description{
Named list mapping transcriptomic index abbreviations to full names.
}
\keyword{datasets}
