% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopExpressionGenes.R
\name{TopExpressionGenes}
\alias{TopExpressionGenes}
\title{Top quantile expression genes}
\usage{
TopExpressionGenes(ExpressionSet, p = 0.99)
}
\arguments{
\item{ExpressionSet}{A standard ExpressionSet}

\item{p}{The quantile probability. Default is \code{p = .99}}
}
\value{
a character vector containing the gene ids residing in the top expression quantile
}
\description{
Select genes residing in the top quantile according to the mean of their expression across the stages
}
\examples{
# reading a standard PhyloExpressionSet
data(PhyloExpressionSetExample)

# select genes with highest variance (top 2\%)  
genes.top_expression <- TopExpressionGenes(PhyloExpressionSetExample, p=.98)

# remove top genes from the PhyloExpressionSet
PhyloExpressionSet.top_removed <- subset(PhyloExpressionSetExample, !(GeneID \%in\% genes.top_expression))

# plot TAI of set with removed quantile
PlotSignature(ExpressionSet = PhyloExpressionSet.top_removed,
                measure       = "TAI", 
                TestStatistic = "FlatLineTest",
                xlab          = "Ontogeny", 
                ylab          = "TAI" )
}
