% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopVarianceGenes.R
\name{TopVarianceGenes}
\alias{TopVarianceGenes}
\title{Top quantile expression-variance genes}
\usage{
TopVarianceGenes(ExpressionSet, p = 0.99)
}
\arguments{
\item{ExpressionSet}{A standard ExpressionSet}

\item{p}{The quantile probability. Default is \code{p = .99}}
}
\value{
a character vector containing the gene ids residing in the top variance expression quantile
}
\description{
Select genes residing in the top quantile according to the variance of their expression across the stages
}
\examples{
# reading a standard PhyloExpressionSet
data(PhyloExpressionSetExample)

# select genes with highest variance (top 5\%)  
genes.top_variance <- TopVarianceGenes(PhyloExpressionSetExample, p=.95)

}
\author{
Stefan Manolache
}
