% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_transforms.R
\name{transform_counts}
\alias{transform_counts}
\title{Transform Expression Counts in PhyloExpressionSet}
\usage{
transform_counts(
  phyex_set,
  FUN,
  FUN_name = deparse(substitute(FUN)),
  new_name = NULL,
  ...
)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{FUN}{Function to apply}

\item{FUN_name}{Name of the transformation function (optional)}

\item{new_name}{Name for the new dataset (optional)}

\item{...}{Additional arguments passed to FUN}
}
\value{
A PhyloExpressionSet object with transformed expression data
}
\description{
Apply a transformation function to the expression counts in a PhyloExpressionSet.
}
