% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_bulk.R
\name{TXI_std_dev}
\alias{TXI_std_dev}
\title{Standard Deviation for TXI}
\usage{
TXI_std_dev(phyex_set)
}
\arguments{
\item{phyex_set}{A BulkPhyloExpressionSet object}
}
\value{
Named numeric vector of standard deviations, names are identities
}
\description{
Return a named vector of standard deviations for the TXI for each identity.
}
