% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_bulk.R
\name{as_BulkPhyloExpressionSet}
\alias{as_BulkPhyloExpressionSet}
\title{as_BulkPhyloExpressionSet}
\usage{
as_BulkPhyloExpressionSet(
  data,
  groups = colnames(data[, 3:ncol(data)]),
  name = deparse(substitute(data)),
  strata_legend = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame with phylostratum assignments and gene expression data}

\item{groups}{Vector of group labels for the samples/replicates}

\item{name}{Character string to name the dataset}

\item{strata_legend}{Optional data frame mapping phylostratum numbers to labels}

\item{...}{Additional arguments passed to BulkPhyloExpressionSet_from_df}
}
\value{
A BulkPhyloExpressionSet object
}
\description{
This function is an alias for \code{BulkPhyloExpressionSet_from_df}. 
Please refer to the documentation for \code{BulkPhyloExpressionSet_from_df} for usage details, arguments, and examples.
}
\seealso{
\code{\link{BulkPhyloExpressionSet_from_df}}
}
