% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{bar.colors}
\alias{bar.colors}
\title{Color palette for barplots}
\usage{
bar.colors(n)
}
\arguments{
\item{n}{the number of colors to be in the palette.}
}
\value{
a character vector containing different color names that can be used for barplots.
}
\description{
A nice color palette for barplots with several bars.
}
\details{
This function can be used to select colors for bar plots.
}
\examples{

# get 5 different colors for 5 different bars
barplot_colors <- bar.colors(5)
}
\author{
Hajk-Georg Drost
}
