% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiffGenes.R
\name{DiffGenes}
\alias{DiffGenes}
\title{Differential Gene Expression Analysis}
\usage{
DiffGenes(ExpressionSet, nrep, method = "foldchange",
  p.adjust.method = NULL, stage.names = NULL)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{nrep}{either a numeric value specifying the constant number of replicates per stage or a numeric vector specifying the variable number of replicates for each stage position.}

\item{method}{method to detect differentially expressed genes.}

\item{p.adjust.method}{p value correction method.}

\item{stage.names}{a character vector specifying the new names of collapsed stages.}
}
\description{
Detect differentially expressed genes (DEGs) in a standard \code{ExpressionSet} object.
}
\details{
Available methods for the detection of differentially expressed genes:

\itemize{
\item \code{method = "foldchange"}: ratio of replicate means between developmental stages.
\item \code{method = "log-foldchange"}: difference of replicate log-means between developmental stages.
\item \code{method = "t.test"}: Welch t.test between replicate expression levels between two samples.
}
}
\examples{
data(PhyloExpressionSetExample)

# Detection of DEGs using the fold-change measure
DEGs <- DiffGenes(ExpressionSet = PhyloExpressionSetExample[1:5,1:8],
                  nrep          = 2,
                  method        = "foldchange",
                  stage.names   = c("S1","S2","S3"))


head(DEGs)
}
\author{
Hajk-Georg Drost
}

