% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_bulk.R
\name{.collapse_replicates}
\alias{.collapse_replicates}
\title{Collapse Expression Data Across Replicates}
\usage{
.collapse_replicates(expression, groups)
}
\arguments{
\item{expression}{Matrix of expression counts}

\item{groups}{Factor indicating group membership}
}
\value{
Matrix with collapsed expression data
}
\description{
Internal function to collapse expression data across replicates by taking row means.
}
\keyword{internal}
