% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.pseudobulk_expression}
\alias{.pseudobulk_expression}
\title{Create Pseudobulk Expression Data}
\usage{
.pseudobulk_expression(expression, groups)
}
\arguments{
\item{expression}{Expression matrix with genes as rows and cells as columns}

\item{groups}{Factor vector indicating which group each cell belongs to}
}
\value{
Matrix of pseudobulked expression
}
\description{
Aggregate single-cell data by cell identity to create pseudobulk expression.
}
\keyword{internal}
