% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_sc.R
\name{.pseudobulk_expression}
\alias{.pseudobulk_expression}
\title{Create Pseudobulk Expression Data}
\usage{
.pseudobulk_expression(seurat, layer)
}
\arguments{
\item{seurat}{A Seurat object}

\item{layer}{Expression layer to use}
}
\value{
Matrix of pseudobulked expression
}
\description{
Aggregate single-cell data by cell identity (using Seurat Idents) to create pseudobulk expression.
}
\keyword{internal}
