% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_phyex_set.R
\docType{data}
\name{example_phyex_set}
\alias{example_phyex_set}
\title{Example phyex set}
\format{
A PhyloExpressionObject with 8 developmental stages (3 reps each) and 27520 genes
}
\source{
phyexSets::Athaliana.embryogenesis_2019 matched with the Arabidopsis thaliana phylomap from phylomapr
}
\usage{
example_phyex_set
}
\description{
Arabidopsis thaliana embryogenesis dataset from Hoffman et al. 2019. check the phyexSet package for more detials
}
\keyword{datasets}
