% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_phyex_set_sc}
\alias{example_phyex_set_sc}
\title{Load Example Single-Cell PhyloExpressionSet}
\format{
A ScPhyloExpressionSet with 1000 cells and 1000 genes
}
\source{
Synthetically generated data
}
\usage{
example_phyex_set_sc
}
\description{
Creates and returns an example ScPhyloExpressionSet object for testing and examples.
}
\keyword{datasets}
