% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test_result.R
\name{exp_p}
\alias{exp_p}
\title{Format P-Value for Scientific Notation}
\usage{
exp_p(p)
}
\arguments{
\item{p}{Numeric p-value}
}
\value{
Expression object for use in plot annotations
}
\description{
Format p-values in scientific notation for plot annotations.
}
\examples{
# Format p-value for plotting
# expr <- exp_p(0.001)

}
\keyword{internal}
