% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_filter.R
\name{genes_top_mean}
\alias{genes_top_mean}
\title{Select Top Mean Expressed Genes}
\usage{
genes_top_mean(phyex_set, p = 0.99)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{p}{Quantile threshold for gene selection (default: 0.99)}
}
\value{
Character vector of gene IDs with mean expression >= p quantile
}
\description{
Select genes with the highest mean expression across samples.
}
\details{
This function identifies genes with the highest mean expression levels,
which are often the most reliably detected and functionally important.
}
\examples{
# Select top 1\% most expressed genes by mean
# high_expr_genes <- genes_top_mean(phyex_set, p = 0.99)

}
