% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_filter.R
\name{genes_top_variance}
\alias{genes_top_variance}
\title{Select Top Variable Genes}
\usage{
genes_top_variance(phyex_set, p = 0.99)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{p}{Quantile threshold for gene selection (default: 0.99)}
}
\value{
Character vector of gene IDs with variance >= p quantile
}
\description{
Select genes with the highest variance across samples.
}
\details{
This function identifies genes with the highest variance across samples,
which are often the most informative for downstream analyses.
}
\examples{
# Select top 1\% most variable genes
high_var_genes <- genes_top_variance(example_phyex_set, p = 0.99)

}
