% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric Mean}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{a numeric vector for which geometric mean computations shall be performed.}
}
\description{
This function computes the geometric mean of a numeric input vector \code{x}.
}
\examples{
\dontshow{if (FALSE ) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- 1:10

geom_mean(x)
\dontshow{\}) # examplesIf}
}
\author{
Hajk-Georg Drost
}
\keyword{internal}
