% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_patterns.R
\name{mod_pi}
\alias{mod_pi}
\title{Modulo Pi Function}
\usage{
mod_pi(x)
}
\arguments{
\item{x}{Numeric vector of angles}
}
\value{
Numeric vector of wrapped angles
}
\description{
Wrap angles to the range (-π, π].
}
\details{
Ensures angles are in the standard range (-π, π] by wrapping around 2π.
}
\keyword{internal}
