% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_expression.R
\name{plot_distribution_expression}
\alias{plot_distribution_expression}
\title{Comparing expression levels distributions across developmental stages}
\usage{
plot_distribution_expression(
  phyex_set,
  show_conditions = TRUE,
  show_strata = FALSE,
  seed = 123
)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object.}

\item{show_conditions}{Logical, whether to show condition-specific distributions.}

\item{show_strata}{Logical, whether to show stratum-specific distributions.}

\item{seed}{Seed for reproducible color selection.}
}
\description{
\code{plot_distribution_expression} generates plots that help to compare the distribution
of expression levels through various developmental stages, highlighting each stage with
distinct colors.
}
\section{Recommendation}{
 
Apply a square root transformation to enhance the visualization of differences
in the distributions: \code{plot_distribution_expression(transform_counts(phyex_set, sqrt))}
}

\author{
Filipa Martins Costa
}
