% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_expression.R
\name{plot_distribution_expression}
\alias{plot_distribution_expression}
\title{Comparing expression levels distributions across the different developmental stages}
\usage{
plot_distribution_expression(
  phyex_set,
  show_conditions = TRUE,
  show_strata = FALSE,
  seed = 123
)
}
\arguments{
\item{seed}{defines the colors for the different developmetal stages}

\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{stages}{a numeric vector specifying the indices of the stages to compare. Each index
corresponds to a stage in the ExpressionSet. Starts in one.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{figure title.}
}
\description{
\emph{plot_distribution_partialTAI} generates 2 plots that help to compare the distribution
of expression levels through various developmental stages, highlighting each stage with
distinct colors.
}
\author{
Filipa Martins Costa
}
