% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_expression.R
\name{plot_distribution_expression}
\alias{plot_distribution_expression}
\title{Comparing expression levels distributions across stages}
\usage{
plot_distribution_expression(
  ExpressionSet,
  stages = 1:(ncol(ExpressionSet) - 2),
  xlab = "Expression",
  ylab = "Density",
  main = "Density Distribution of Expression by Developmental Stage",
  seed = 123
)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{stages}{a numeric vector specifying the indices of the stages to compare. Each index 
corresponds to a stage in the ExpressionSet. Starts in one.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{figure title.}

\item{seed}{defines the colors for the different developmental stages}
}
\description{
\emph{plot_distribution_partialTAI} generates two plots that help to compare the distribution
of expression levels through various developmental stages, highlighting each stage with 
distinct colors.
}
\details{
Recommendation - Apply a square root transformation to enhance the visualization of differences 
in the distributions: \code{plot_distribution_partialTAI(tf(ExpressionSet, sqrt))}
}
\author{
Filipa Martins Costa
}
