% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotGeneSet.R
\name{PlotGeneSet}
\alias{PlotGeneSet}
\title{Plot the Expression Profiles of a Gene Set}
\usage{
PlotGeneSet(ExpressionSet, gene.set, get.subset = FALSE, colors = NULL, ...)
}
\arguments{
\item{ExpressionSet}{}

\item{gene.set}{}

\item{get.subset}{}

\item{colors}{}

\item{...}{additional parameters passed to \code{\link{matplot}}.}
}
\description{
This function simply visualizes the gene expression profiles of
a defined subset of genes stored in the input \code{ExpressionSet}.
}
\examples{
data(PhyloExpressioNSetExample)
}
\author{
Hajk-Georg Drost
}

