% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSignature.R
\name{PlotSignature}
\alias{PlotSignature}
\title{Plot evolutionary signatures across transcriptomes}
\usage{
PlotSignature(ExpressionSet, measure = "TAI",
  TestStatistic = "FlatLineTest", modules = NULL, permutations = 1000,
  xlab = "Ontogeny", ylab = "Transcriptome Index", main = "", lwd = 4,
  alpha = 0.1, y.ticks = 10)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}

\item{TestStatistic}{a string defining the type of test statistics to be used to quantify the statistical significance the present phylotranscriptomics pattern.
Possible values can be: \code{TestStatistic} = \code{"FlatLineTest"} : Statistical test for the deviation from a flat line.
\code{TestStatistic} = \code{"ReductiveHourglassTest"} : Statistical test for the existence of a hourglass shape (high-low-high pattern).
\code{TestStatistic} = \code{"EarlyConservationTest"} : Statistical test for the existence of a earlyconservation pattern (low-high-high pattern).}

\item{modules}{a list storing three elements for the \code{\link{ReductiveHourglassTest}} or \code{\link{EarlyConservationTest}}: early, mid, and late. 
Each element expects a numeric vector specifying the developmental stages 
or experiments that correspond to each module. For example, 
\code{module} = \code{list(early = 1:2, mid = 3:5, late = 6:7)} devides a dataset storing seven developmental stages into 3 modules.}

\item{permutations}{a numeric value specifying the number of permutations to be performed for the \code{\link{FlatLineTest}}, \code{\link{EarlyConservationTest}} or \code{\link{ReductiveHourglassTest}}.}
}
\description{
E
}
\author{
Hajk-Georg Drost
}

