% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloset_base.R
\name{pTXI}
\alias{pTXI}
\title{Calculate Phylostratum-Specific Transcriptomic Index}
\usage{
pTXI(phyex_set, reps = FALSE)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}
}
\value{
Matrix of pTXI values
}
\description{
Calculate pTXI values for expression data. This is a generic function
that dispatches based on the input type.
}
