% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TDI}
\alias{TDI}
\title{Function to compute the Transcriptome Divergence Index (TDI)}
\usage{
TDI(DivergenceExpressionSet)
}
\arguments{
\item{DivergenceExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}
}
\value{
a numeric vector containing the TDI values for all given developmental stages.
}
\description{
This function computes the sequence distance based transcriptome divergence index (TDI) introduced by
Quint et al., 2012.
}
\details{
The TDI measure represents the weighted arithmetic mean (expression levels as
weights for the divergence-stratum value) over all gene divergence categories denoted as \emph{divergence-strata}.


\deqn{TDI_s = \sum (e_is * ds_i) / \sum e_is}

where TDI_s denotes the TDI value in developmental stage s, e_is denotes the gene expression level of gene i in stage s, and ds_i denotes the corresponding divergence-stratum of gene i, \eqn{i = 1,...,N} and N = total number of genes.

Internally the function is written in C++ to speed up TDI computations.
}
\examples{
# reading a standard DivergenceExpressionSet
data(DivergenceExpressionSetExample)

# computing the TDI profile of a given DivergenceExpressionSet object
TDIs <- TDI(DivergenceExpressionSetExample)
}
\author{
Hajk-Georg Drost
}
\references{
Quint M et al. 2012. "A transcriptomic hourglass in plant embryogenesis". Nature (490): 98-101.

Drost et al. (2014), Active maintenance of phylotranscriptomic hourglass patterns in animal and plant embryogenesis.
}
\seealso{
\code{\link{TAI}},  \code{\link{PlotPattern}}, \code{\link{FlatLineTest}}, \code{\link{ReductiveHourglassTest}}
}

