% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDI.R
\name{TDI}
\alias{TDI}
\title{Compute the Transcriptome Divergence Index (TDI)}
\usage{
TDI(DivergenceExpressionSet)
}
\arguments{
\item{DivergenceExpressionSet}{a standard PhyloExpressionSet or DivergenceExpressionSet object.}
}
\value{
a numeric vector containing the TDI values for all given developmental stages.
}
\description{
This function computes the sequence distance based transcriptome divergence index (TDI) introduced by
Quint et al., 2012.
}
\details{
The TDI measure represents the weighted arithmetic mean (expression levels as
weights for the divergence-stratum value) over all gene divergence categories denoted as \emph{divergence-strata}.


\deqn{TDI_s = \sum (e_is * ds_i) / \sum e_is}

where TDI_s denotes the TDI value in developmental stage s, e_is denotes the gene expression level of gene i in stage s, and ds_i denotes the corresponding divergence-stratum of gene i, \eqn{i = 1,...,N} and N = total number of genes.

Internally the function is written in C++ to speed up TDI computations.
}
\examples{

# reading a standard DivergenceExpressionSet
data(DivergenceExpressionSetExample)

# computing the TDI profile of a given DivergenceExpressionSet object
TDIs <- TDI(DivergenceExpressionSetExample)


}
\references{
Quint M et al. (2012). \emph{A transcriptomic hourglass in plant embryogenesis}. Nature (490): 98-101.

Drost HG et al. (2015) Mol Biol Evol. 32 (5): 1221-1231 doi:10.1093/molbev/msv012
}
\seealso{
\code{\link{TAI}},  \code{\link{PlotPattern}}, \code{\link{FlatLineTest}}, \code{\link{ReductiveHourglassTest}}
}
\author{
Hajk-Georg Drost
}
