% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEI.R
\name{TEI}
\alias{TEI}
\title{Compute the Transcriptome Evolutionary Index (TEI)}
\usage{
TEI(
  ExpressionSet,
  Phylostratum = NULL,
  split = 1e+05,
  showprogress = TRUE,
  threads = 1
)
}
\arguments{
\item{ExpressionSet}{expression object with rownames as GeneID (dgCMatrix)
or standard PhyloExpressionSet object.}

\item{Phylostratum}{a named vector representing phylostratum per GeneID with
names as GeneID (not used if Expression is PhyloExpressionSet).}

\item{split}{specify number of columns to split}

\item{showprogress}{boolean if progressbar should be shown}

\item{threads}{specify number of threads}
}
\value{
a numeric vector containing the TEI values for all given cells or
developmental stages.
}
\description{
This function computes the phylogenetically based transcriptome evolutionary
index (TEI) similar to Domazet-Loso & Tautz, 2010.
}
\details{
The TEI measure represents the weighted arithmetic mean
(expression levels as weights for the phylostratum value) over all
evolutionary age categories denoted as \emph{phylostra}.

\deqn{TEI_s = \sum (e_is * ps_i) / \sum e_is}

where TEI_s denotes the TEI value in developmental stage s,
e_is denotes the gene expression level of gene i in stage s,
and ps_i denotes the corresponding phylostratum of gene i,
\eqn{i = 1,...,N} and N = total number of genes.
}
\examples{

# reading a standard PhyloExpressionSet
data(PhyloExpressionSetExample, package = "myTAI")

# computing the TEI profile of a given PhyloExpressionSet object
TEI <- TEI(PhyloExpressionSetExample)

}
\references{
Domazet-Loso T. and Tautz D. (2010).
\emph{A phylogenetically based transcriptome age index mirrors ontogenetic
divergence patterns}. Nature (468): 815-818.

Quint M et al. (2012).
\emph{A transcriptomic hourglass in plant embryogenesis}.
Nature (490): 98-101.

Drost HG et al. (2015)
Mol Biol Evol. 32 (5): 1221-1231 doi:10.1093/molbev/msv012
}
\author{
Kristian K Ullrich
}
