% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_transforms.R
\name{tf}
\alias{tf}
\title{Short Alias for Transform Counts}
\usage{
tf(phyex_set, FUN, FUN_name = deparse(substitute(FUN)), new_name = NULL, ...)
}
\description{
Convenience alias for transform_counts function.
}
