% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_transforms.R
\name{tf}
\alias{tf}
\title{Short Alias for Transform Counts}
\usage{
tf(phyex_set, ...)
}
\arguments{
\item{phyex_set}{A PhyloExpressionSet object}

\item{...}{Arguments passed to transform_counts}
}
\value{
A PhyloExpressionSet object with transformed expression data
}
\description{
Convenience alias for transform_counts function.
}
\examples{
# Short alias for transformation
# log_set <- tf(phyex_set, log1p)

}
