% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_stability.R
\name{tf_stability}
\alias{tf_stability}
\title{Perform Permutation Tests Under Different Transformations}
\usage{
tf_stability(
  phyex_set,
  TestStatistic,
  transforms = c("none", "sqrt", "log2", "rank", "squared"),
  modules = NULL,
  permutations = 1000,
  pseudocount = 1
)
}
\arguments{
\item{phyex_set}{a \code{PhyloExpressionSet} S7 object.}

\item{TestStatistic}{a conservation test function}

\item{transforms}{character vector of transformation names (e.g., "log2", "sqrt", "none", "rank", "squared").}

\item{modules}{a list with three elements: early, mid, and late. Each is a numeric vector specifying developmental stages or experiments for each module.}

\item{permutations}{number of permutations for the test.}

\item{pseudocount}{numeric, value to add before log transforms.}
}
\value{
Named numeric vector of p-values for each transformation.
}
\description{
\emph{tfStability} statistically evaluates the
stability of phylotranscriptomics permutation tests (e.g., \code{ReductiveHourglassTest}, \code{FlatLineTest}, etc.)
under different data transformations using a \code{PhyloExpressionSet} S7 object.
}
\details{
Assesses the stability of data transforms on the permutation test of choice.
See \code{\link{tf}}, \code{\link{ReductiveHourglassTest}}, \code{\link{FlatLineTest}}, etc.
}
\references{
Lotharukpong JS et al. (2023) (unpublished)
}
\author{
Jaruwatana Sodai Lotharukpong
}
