% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfPS.R
\name{tfPS}
\alias{tfPS}
\title{Transform Phylostratum Values}
\usage{
tfPS(ExpressionSet, transform)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet object.}

\item{transform}{a character vector of any valid function that transforms PS values.
Possible values can be:
\itemize{
\item \code{transform} = \code{"qr"} (or \code{"quantilerank"}) : 
quantile rank transformation analogous to Julia function \code{StatsBase.quantilerank} 
using \code{method = :tied}.
}}
}
\value{
a standard PhloExpressionSet object storing transformed Phylostratum levels.
}
\description{
This function performs transformation of phylostratum values.
}
\details{
This function transforms the phylostratum assignment. 
The return value of this function is a PhyloExpressionSet object with 
transformed phylostratum \code{tfPhylostratum} as the first column, satisfying
\code{\link{is.ExpressionSet}}. Note that the input \code{transform} must be an
available function, currently limited to only \code{"qr"} (or \code{"quantilerank"}).
}
\examples{
# source the example dataset
data(PhyloExpressionSetExample)
 
# get the relative expression profiles for each phylostratum
tfPES <- tfPS(PhyloExpressionSetExample, transform = "qr")
head(tfPES)

}
\seealso{
\code{\link{tf}}
}
\author{
Jaruwatana Sodai Lotharukpong and Lukas Maischak
}
